requirements.txt
streamlit
google-generativeai
gTTS
SpeechRecognition
pydub
